+++ processor/src/main/java/react4j/processor/Constants.java
@@ -2,6 +2,9 @@ package react4j.processor;
 
 final class Constants
 {
+  static final String DAGGER_SUBCOMPONENT_CLASSNAME = "dagger.Subcomponent";
+  static final String DAGGER_MODULE_CLASSNAME = "dagger.Module";
+  static final String DAGGER_PROVIDES_CLASSNAME = "dagger.Provides";
   static final String JS_FUNCTION_CLASSNAME = "jsinterop.annotations.JsFunction";
   static final String COMPONENT_CLASSNAME = "react4j.core.Component";
   static final String PROCEDURE_CLASSNAME = "react4j.core.Procedure";