+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+##### Changed
+* Incorporate the processorpath buildr addon to simplify build configuration.
+
 ### [v0.21](https://github.com/react4j/react4j/tree/v0.21) (2017-12-04)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.20...v0.21)